refactor(redis): separate test modules into files
also move helpers into tests/common
